setup to test optimisation of hierarchical factors
