+++ test_autofit/graphical/hierarchical/test_optimise.py
@@ -1,0 +1,15 @@
+import autofit as af
+
+
+def test_optimise(
+        hierarchical_factor
+):
+    optimizer = af.DynestyStatic(
+        maxcall=10
+    )
+    factor = hierarchical_factor.factors[0]
+
+    optimizer.optimise(
+        factor,
+        factor.mean_field_approximation()
+    )
